Added electron functions and tests.
I wanted to add these tests to show that the actual Electron window that
karma is running in is the current electron window. I considered adding
tests to open a new Electron window but figured I could just write a
blog post and/or create a example cljs Electron app repo that people can
reference.